Implement lazyNode with pre/post and change exec to run
